Fix resource load order
